requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
django==4.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
idna==3.10
    # via requests
pyjwt==2.9.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via asgiref
urllib3==2.2.3
    # via requests
